requirements.txt
fastapi[all]
sqlalchemy[asyncio]
asyncpg==0.31.0
psycopg2-binary==2.9.10
alembic
pydantic-settings
celery
redis
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
httpx
pyyaml
python-dotenv
boto3
pypdf
openai
langchain
google-genai
slowapi
cloudinary
groq
aiohttp
gunicorn
uvicorn
stripe
mangum
